+++ lightly_studio/src/lightly_studio/resolvers/image_resolver_new/__init__.py
@@ -1,0 +1,8 @@
+"""Resolvers for database operations."""
+
+from lightly_studio.resolvers.image_resolver_new.create import create, create_many
+
+__all__ = [
+    "create",
+    "create_many",
+]
