+++ lightly_studio/tests/helpers_resolvers.py
@@ -37,7 +37,7 @@ from lightly_studio.resolvers import (
     annotation_resolver,
     dataset_resolver,
     embedding_model_resolver,
-    image_resolver,
+    image_resolver_new,
     sample_embedding_resolver,
     tag_resolver,
 )
@@ -89,7 +89,7 @@ def create_image(
     height: int = 1080,
 ) -> ImageTable:
     """Helper function to create a sample."""
-    return image_resolver.create(
+    return image_resolver_new.create(
         session=session,
         sample=ImageCreate(
             dataset_id=dataset_id,
