+++ lightly_studio_view/src/lib/schema.d.ts
@@ -1528,6 +1528,7 @@ export interface components {
             name: string;
             /** Parent Dataset Id */
             parent_dataset_id?: string | null;
+            sample_type: components["schemas"]["SampleType"];
         };
         /**
          * DatasetTable
@@ -1538,6 +1539,7 @@ export interface components {
             name: string;
             /** Parent Dataset Id */
             parent_dataset_id?: string | null;
+            sample_type: components["schemas"]["SampleType"];
             /**
              * Dataset Id
              * Format: uuid
@@ -1563,6 +1565,7 @@ export interface components {
             name: string;
             /** Parent Dataset Id */
             parent_dataset_id?: string | null;
+            sample_type: components["schemas"]["SampleType"];
             /**
              * Dataset Id
              * Format: uuid
@@ -1588,6 +1591,7 @@ export interface components {
             name: string;
             /** Parent Dataset Id */
             parent_dataset_id?: string | null;
+            sample_type: components["schemas"]["SampleType"];
             /**
              * Dataset Id
              * Format: uuid
@@ -2029,6 +2033,12 @@ export interface components {
              */
             sample_ids?: string[] | null;
         };
+        /**
+         * SampleType
+         * @description The type of samples in the dataset.
+         * @enum {string}
+         */
+        SampleType: "video" | "image" | "image_annotation";
         /**
          * SamplesToRefineResponse
          * @description Response for samples for classifier refinement.
